Update commit script.

+++ scripts/commitxml.ps1
@@ -5,15 +5,37 @@ Get-ChildItem _xml
 
 Write-Output "--------------------------"
 
-Remove-Item _xml -Recurse
+If (Test-Path _xml){
+    Remove-Item _xml -Recurse
+}
+
+If (Test-Path _mt){
+    Remove-Item _mt -Recurse
+}
+
+If (Test-Path docpack){
+    Remove-Item docpack -Recurse
+}
+
+If (Test-Path mdoc.zip){
+    Remove-Item mdoc.zip
+}
+
+If (Test-Path nue.zip){
+    Remove-Item nue.zip
+}
+
+If (Test-Path docpac.zip){
+    Remove-Item docpac.zip
+}
+
+If (Test-Path march-train.zip){
+    Remove-Item march-train.zip
+}
+
 Rename-Item mdoc-output -NewName _xml
+
 Get-ChildItem $srcFolder | Where-Object {$_.PSIsContainer -and ($_ -match '^mdoc.*$')} | Remove-Item -Recurse
-Remove-Item mdoc.zip
-Remove-Item nue.zip
-Remove-Item docpac.zip
-Remove-Item march-train.zip
-Remove-Item _mt
-Remove-Item docpack
 
 # Remove the warning, that otherwise breaks the build
 git config --global core.safecrlf false
